feat: Добавить хук авторизацции
